test: move `pasteFiles` / `dropFiles` into `@meowdown/vitest/file-events`
