+++ packages/react/src/components/editor.test.tsx
@@ -1,5 +1,6 @@
 import '../testing/index.ts'
 
+import { dropFiles } from '@meowdown/vitest/file-events'
 import { pasteText } from '@prosekit/core/test'
 import { TextSelection } from '@prosekit/pm/state'
 import { useEditor } from '@prosekit/react'
@@ -235,22 +236,12 @@ describe('MeowdownEditor', () => {
     )
     await expect.element(screen.getByText('Drop zone')).toBeInTheDocument()
 
+    const view = ref.current?.editor?.view
+    if (!view) throw new Error('editor not mounted')
     // Simulate a Finder drop: a synthetic `drop` carrying an image File, aimed
     // at the paragraph so ProseKit's drop-indicator also claims the position.
-    const target = pmRoot.getByText('Drop zone').element()
-    const rect = target.getBoundingClientRect()
     const file = new File(['cat'], 'cat.png', { type: 'image/png' })
-    const dataTransfer = new DataTransfer()
-    dataTransfer.items.add(file)
-    pmRoot.element().dispatchEvent(
-      new DragEvent('drop', {
-        dataTransfer,
-        clientX: rect.left + rect.width / 2,
-        clientY: rect.top + rect.height / 2,
-        bubbles: true,
-        cancelable: true,
-      }),
-    )
+    dropFiles(view, [file], 1)
 
     await vi.waitFor(() => expect(onFilePaste).toHaveBeenCalledWith(file))
     await vi.waitFor(() => expect(ref.current?.getMarkdown()).toContain('![](https://cdn/cat.png)'))
